+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog:
 
-## [0.1.0] 2023-02-21
+## [0.1.1] 18-03-2023
+
+- Added preferences menu.
+	- Log messages can be toggled.
+	- Hierarchy title can be toggled.
+	- Hierarchy prefix can be changed (Must be 3 or more characters long).
+	
+- Cleaned up editor pref loading & saving.
+- Bug fixes.
+
+## [0.1.0] 21-02-2023
 
 - Initial public version